Moved code around to get rid of extra sound_generator file. Renamed DatasetCreator to be clearer
